pyproject.toml
[tool.poetry]
name = "services"
version = "0.1.0"
description = ""
authors = ["Prajwal Poojary <pprajwal312@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
azure-identity = "^1.19.0"
python-dotenv = "^1.0.1"
azure-mgmt-logic = "^10.0.0"
azure-storage-blob = "^12.24.1"
azure-keyvault = "^4.2.0"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
